Added API endpoints for update, delete, detail for products

+++ src/products/products.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query, BadRequestException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -6,6 +6,8 @@ import { Role } from 'src/users/enum/role.enum';
 import { Roles } from 'src/auth/decorator/roles.decorator';
 import { plainToInstance } from 'class-transformer';
 import { PageRequestDTO } from './dto/pagination/page.request.dto';
+import { Product } from './entities/product.entity';
+import { PageResponseDTO } from './dto/pagination/page.response.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -13,30 +15,34 @@ export class ProductsController {
 
   @Roles(Role.Admin)
   @Post()
-  create(@Body() createProductDto: CreateProductDto) {
-    return this.productsService.create(createProductDto);
+  async create(@Body() createProductDto: CreateProductDto): Promise<Product> {
+    return await this.productsService.create(createProductDto);
   }
 
   @Get()
-  findAll(@Query() pageRequestDTO: PageRequestDTO) {
+  async findAll(@Query() pageRequestDTO: PageRequestDTO): Promise<PageResponseDTO<Product>> {
     const requestDTO = plainToInstance(PageRequestDTO, pageRequestDTO);
-    return this.productsService.findAll(requestDTO);
+    return await this.productsService.findAll(requestDTO);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    console.log('here i am in products controller');
-    return this.productsService.findOne(+id);
+  async findOne(@Param('id') id: string): Promise<Product> {
+    return await this.productsService.findOne(+id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
-    return this.productsService.update(+id, updateProductDto);
+  async update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto): Promise<Product> {
+    if (isNaN(+id)) {
+      throw new BadRequestException('Invalid product id');
+    }
+    return await this.productsService.update(+id, updateProductDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (isNaN(+id)) {
+      throw new BadRequestException('Invalid product id');
+    }
     return this.productsService.remove(+id);
   }
-
 }
